# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorCode": {
        "description": "The stable wire-level failure codes currently emitted by the control plane.",
        "enum": [
          "ADMINISTRATION_WOULD_BE_LOST",
          "LOGIN_NAME_TAKEN",
          "PASSWORD_TOO_SHORT",
          "PERMISSION_UNREGISTERED",
          "ROLE_CODE_TAKEN",
          "ROLE_NOT_FOUND",
          "USER_NOT_FOUND",
          "ACCOUNT_DEACTIVATED",
          "AUTHENTICATION_REQUIRED",
          "CREDENTIALS_REJECTED",
          "CSRF_TOKEN_INVALID",
          "INFERENCE_BACKEND_ENDPOINT_TAKEN",
          "INFERENCE_BACKEND_NOT_FOUND",
          "INFERENCE_HOST_DEACTIVATED",
          "INFERENCE_HOST_HAS_BACKENDS",
          "INFERENCE_HOST_NAME_TAKEN",
          "INFERENCE_HOST_NOT_FOUND",
          "INTERNAL_ERROR",
          "PERMISSION_DENIED",
          "REQUEST_INVALID",
          "SESSION_INVALID",
          "STALE_REVISION"
        ],
        "title": "ApiErrorCode",
        "type": "string"
      },
...
      "DeviceStatus": {
        "description": "Whether a configurable device still takes part in new bindings and operation.\n\nAn enum rather than an `is_active` boolean: `CONTEXT.md` gives 停用 its own definition\nacross every mutable configuration object. Deactivation is reversible and never\ncascades — a deactivated host's backends keep their rows and their history.",
        "enum": [
          "active",
          "deactivated"
        ],
        "title": "DeviceStatus",
        "type": "string"
      },
...
      "FieldError": {
        "description": "One rejected input, named so the Web form can put the message beside it.",
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "HostConfiguration": {
        "description": "The host's whole editable configuration, exactly as the edit form submits it.\n\n§5.19: durations and thresholds are configured values, and 0 or negative — or a\nwatermark of 0 or 100 — is a configuration error refused at save time. A URL with an\nembedded `user:password` is refused at the contract (ADR-0008), before it can be stored\nor echoed back.",
        "properties": {
          "address": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Address",
            "type": "string"
          },
          "disk_watermark_percent": {
            "maximum": 99.0,
            "minimum": 1.0,
            "title": "Disk Watermark Percent",
            "type": "integer"
          },
          "mediamtx_address": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "pattern": "^https?://",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mediamtx Address"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "recording_window_seconds": {
            "exclusiveMinimum": 0.0,
            "title": "Recording Window Seconds",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "address",
          "recording_window_seconds",
          "disk_watermark_percent"
        ],
        "title": "HostConfiguration",
        "type": "object"
      },
      "HostStatus": {
        "description": "The one field 停用 and 恢复 toggle, as the two values of one subresource.",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/DeviceStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "HostStatus",
        "type": "object"
      },
...
      "InferenceHostView": {
        "description": "One host as the API carries it. `revision` is what If-Match echoes.",
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "format": "uuid",
            "title": "Created By",
            "type": "string"
          },
          "disk_watermark_percent": {
            "title": "Disk Watermark Percent",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mediamtx_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mediamtx Address"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "recording_window_seconds": {
            "title": "Recording Window Seconds",
            "type": "integer"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/DeviceStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "format": "uuid",
            "title": "Updated By",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "mediamtx_address",
          "recording_window_seconds",
          "disk_watermark_percent",
          "status",
          "revision",
          "created_by",
          "updated_by",
          "created_at",
          "updated_at"
        ],
        "title": "InferenceHostView",
        "type": "object"
      },
...
      "ItemPage_InferenceHostView_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InferenceHostView"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "ItemPage[InferenceHostView]",
        "type": "object"
      },
...
      "ProblemDocument": {
        "description": "RFC 9457 plus the stable control-plane extensions from §5.15.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "error_code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "field_errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Errors"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "const": "about:blank",
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "error_code"
        ],
        "title": "ProblemDocument",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SOP compliance center backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/inference-hosts": {
      "get": {
        "description": "One page of hosts, newest first, under §5.15's envelope.",
        "operationId": "listInferenceHosts",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemPage_InferenceHostView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Authentication required or session invalid"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Permission denied or CSRF token invalid"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Request invalid"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "List The Hosts",
        "tags": [
          "device"
        ],
        "x-required-permission": "device.inference_host.view"
      },
      "post": {
        "description": "Register a physical inference machine.",
        "operationId": "createInferenceHost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostConfiguration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceHostView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Authentication required or session invalid"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Permission denied or CSRF token invalid"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Host name already taken"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Request invalid"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Create A Host",
        "tags": [
          "device"
        ],
        "x-required-permission": "device.inference_host.edit"
      }
    },
    "/api/v1/inference-hosts/{host_id}": {
      "delete": {
        "description": "Delete the host outright — the irreversible operation 停用 exists to avoid.",
        "operationId": "deleteInferenceHost",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Host Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Authentication required or session invalid"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Permission denied or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Host not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Host still carries backends"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Request invalid"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Delete A Host",
        "tags": [
          "device"
        ],
        "x-required-permission": "device.inference_host.delete"
      },
      "get": {
        "description": "Read one host.",
        "operationId": "readInferenceHost",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Host Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceHostView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Authentication required or session invalid"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Permission denied or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Host not found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Request invalid"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Read A Host",
        "tags": [
          "device"
        ],
        "x-required-permission": "device.inference_host.view"
      },
      "patch": {
        "description": "Replace the host's whole configuration.\n\nThe request must carry `If-Match: <revision>` — the revision the caller read. A body\nwithout it cannot say what it believed it was editing, so it is refused rather than\nallowed to overwrite blind.",
        "operationId": "editInferenceHost",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Host Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostConfiguration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceHostView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Authentication required or session invalid"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Permission denied or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Host not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Revision moved (STALE_REVISION), or the name is taken"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Request invalid"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Edit A Host",
        "tags": [
          "device"
        ],
        "x-required-permission": "device.inference_host.edit"
      }
    },
    "/api/v1/inference-hosts/{host_id}/status": {
      "put": {
        "description": "停用 or 恢复 the host — the two values of one field, on one subresource.\n\nBoth directions are reversible and idempotent, which is why they share a route; which use\ncase runs is the requested value's, and the match is exhaustive over the enum.",
        "operationId": "setInferenceHostStatus",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Host Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostStatus"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceHostView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Authentication required or session invalid"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Permission denied or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Host not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Revision moved (STALE_REVISION)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Request invalid"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Set The Host Status",
        "tags": [
          "device"
        ],
        "x-required-permission": "device.inference_host.edit"
      }
    },
...
  }
}